--- src/main/java/roomescape/reservation/dto/ReservationRequest.java
@@ -1,11 +1,0 @@
-package roomescape.reservation.dto;
-
-import jakarta.validation.constraints.NotNull;
-import java.time.LocalDate;
-
-public record ReservationRequest(
-        @NotNull LocalDate date,
-        @NotNull Long timeId,
-        @NotNull Long themeId
-) {
-}
