+++ src/main/java/roomescape/reservation/reservation/AdminReservationController.java
@@ -1,8 +1,10 @@
-package roomescape.reservation;
+package roomescape.reservation.reservation;
 
 import jakarta.validation.Valid;
+
 import java.net.URI;
 import java.util.List;
+
 import lombok.AllArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -11,9 +13,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.reservation.dto.AdminFilterReservationRequest;
-import roomescape.reservation.dto.AdminReservationRequest;
-import roomescape.reservation.dto.ReservationResponse;
+import roomescape.reservation.reservation.dto.AdminFilterReservationRequest;
+import roomescape.reservation.reservation.dto.AdminReservationRequest;
+import roomescape.reservation.reservation.dto.ReservationResponse;
 
 @RestController
 @RequestMapping("/admin/reservations")
