+++ src/main/java/roomescape/reservation/reservation/dto/AdminReservationRequest.java
@@ -1,0 +1,13 @@
+package roomescape.reservation.reservation.dto;
+
+import jakarta.validation.constraints.NotNull;
+
+import java.time.LocalDate;
+
+public record AdminReservationRequest(
+        @NotNull LocalDate date,
+        @NotNull Long themeId,
+        @NotNull Long timeId,
+        @NotNull Long memberId
+) {
+}
