+++ src/main/java/roomescape/reservation/reservation/dto/ReservationResponse.java
@@ -1,9 +1,10 @@
-package roomescape.reservation.dto;
+package roomescape.reservation.reservation.dto;
 
 import java.time.LocalDate;
+
 import roomescape.exception.custom.reason.ResponseInvalidException;
 import roomescape.member.dto.MemberResponse;
-import roomescape.reservation.Reservation;
+import roomescape.reservation.reservation.Reservation;
 import roomescape.reservationtime.dto.ReservationTimeResponse;
 import roomescape.theme.dto.ThemeResponse;
