+++ src/main/java/roomescape/reservation/waiting/AdminWaitingController.java
@@ -1,9 +1,8 @@
-package roomescape.waiting;
+package roomescape.reservation.waiting;
 
 import lombok.AllArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import roomescape.waiting.dto.WaitingResponse;
 
 import java.util.List;
