+++ src/test/java/roomescape/LayeredArchitectureTest.java
@@ -3,13 +3,14 @@ package roomescape;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.lang.reflect.Field;
+
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.jdbc.core.JdbcTemplate;
-import roomescape.reservation.ReservationController;
+import roomescape.reservation.reservation.ReservationController;
 
 @SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
 public class LayeredArchitectureTest {
