+++ src/test/java/roomescape/member/MemberServiceTest.java
@@ -9,7 +9,7 @@ import roomescape.auth.dto.LoginMember;
 import roomescape.exception.custom.reason.member.MemberEmailConflictException;
 import roomescape.member.dto.MemberRequest;
 import roomescape.member.dto.MemberResponse;
-import roomescape.reservation.ReservationRepository;
+import roomescape.reservation.reservation.ReservationRepository;
 import roomescape.util.TestFactory;
 
 import java.util.List;
