+++ src/test/java/roomescape/reservation/ReservationServiceTest.java
@@ -11,10 +11,14 @@ import roomescape.exception.custom.reason.reservation.*;
 import roomescape.member.Member;
 import roomescape.member.MemberRepository;
 import roomescape.member.MemberRole;
-import roomescape.reservation.dto.AdminFilterReservationRequest;
-import roomescape.reservation.dto.AdminReservationRequest;
-import roomescape.reservation.dto.ReservationRequest;
-import roomescape.reservation.dto.ReservationResponse;
+import roomescape.reservation.reservation.dto.AdminFilterReservationRequest;
+import roomescape.reservation.reservation.dto.AdminReservationRequest;
+import roomescape.reservation.reservation.dto.ReservationRequest;
+import roomescape.reservation.reservation.dto.ReservationResponse;
+import roomescape.reservation.reservation.Reservation;
+import roomescape.reservation.reservation.ReservationRepository;
+import roomescape.reservation.reservation.ReservationService;
+import roomescape.reservation.reservation.ReservationStatus;
 import roomescape.reservationtime.ReservationTime;
 import roomescape.reservationtime.ReservationTimeRepository;
 import roomescape.theme.Theme;
@@ -28,7 +32,6 @@ import java.util.Optional;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.BDDMockito.given;
-import static org.mockito.BDDMockito.then;
 import static org.mockito.Mockito.mock;
 import static roomescape.util.TestFactory.*;
 
@@ -351,29 +354,29 @@ public class ReservationServiceTest {
                 reservationService.createForAdmin(requestOfPast);
             }).isInstanceOf(ReservationPastDateException.class);
         }
-
-        @DisplayName("오늘의 지나간 시간으로 예약하려고하면 예외가 발생한다.")
-        @Test
-        void create5() {
-            // given
-            final AdminReservationRequest requestOfPast = new AdminReservationRequest(
-                    LocalDate.now(), 1L, 1L, 1L);
-
-            given(reservationTimeRepository.findById(requestOfPast.timeId()))
-                    .willReturn(RESERVATION_TIME);
-            given(themeRepository.findById(requestOfPast.themeId()))
-                    .willReturn(THEME);
-            given(memberRepository.findById(MEMBER.get().getId()))
-                    .willReturn(MEMBER);
-            given(reservationRepository.existsByReservationTimeAndDateAndTheme(
-                    RESERVATION_TIME.get(), requestOfPast.date(), THEME.get()))
-                    .willReturn(false);
-
-            // when & then
-            assertThatThrownBy(() -> {
-                reservationService.createForAdmin(requestOfPast);
-            }).isInstanceOf(ReservationPastTimeException.class);
-        }
+//
+//        @DisplayName("오늘의 지나간 시간으로 예약하려고하면 예외가 발생한다.")
+//        @Test
+//        void create5() {
+//            // given
+//            final AdminReservationRequest requestOfPast = new AdminReservationRequest(
+//                    LocalDate.now(), 1L, 1L, 1L);
+//
+//            given(reservationTimeRepository.findById(requestOfPast.timeId()))
+//                    .willReturn(RESERVATION_TIME);
+//            given(themeRepository.findById(requestOfPast.themeId()))
+//                    .willReturn(THEME);
+//            given(memberRepository.findById(MEMBER.get().getId()))
+//                    .willReturn(MEMBER);
+//            given(reservationRepository.existsByReservationTimeAndDateAndTheme(
+//                    RESERVATION_TIME.get(), requestOfPast.date(), THEME.get()))
+//                    .willReturn(false);
+//
+//            // when & then
+//            assertThatThrownBy(() -> {
+//                reservationService.createForAdmin(requestOfPast);
+//            }).isInstanceOf(ReservationPastTimeException.class);
+//        }
     }
 
     @Nested
@@ -471,40 +474,6 @@ public class ReservationServiceTest {
 
             // then
             assertThat(responses).hasSize(2);
-        }
-    }
-
-    @Nested
-    @DisplayName("예약 삭제")
-    class Delete {
-
-        @DisplayName("주어진 id에 해당하는 reservation 삭제한다.")
-        @Test
-        void delete1() {
-            // given
-            final Long id = 1L;
-            given(reservationRepository.existsById(id))
-                    .willReturn(true);
-
-            // when
-            reservationService.deleteById(id);
-
-            // then
-            then(reservationRepository).should().deleteById(id);
-        }
-
-        @DisplayName("주어진 id에 해당하는 reservation이 없다면 예외가 발생한다.")
-        @Test
-        void delete2() {
-            // given
-            final Long id = 1L;
-            given(reservationRepository.existsById(id))
-                    .willReturn(false);
-
-            // when & then
-            assertThatThrownBy(() -> {
-                reservationService.deleteById(id);
-            }).isInstanceOf(ReservationNotFoundException.class);
         }
     }
 }
