+++ src/test/java/roomescape/waiting/WaitingServiceTest.java
@@ -10,16 +10,15 @@ import roomescape.exception.custom.reason.auth.AuthorizationException;
 import roomescape.member.Member;
 import roomescape.member.MemberRepository;
 import roomescape.member.MemberRole;
-import roomescape.reservation.Reservation;
-import roomescape.reservation.ReservationRepository;
-import roomescape.reservation.ReservationStatus;
-import roomescape.reservation.dto.ReservationResponse;
+import roomescape.reservation.reservation.Reservation;
+import roomescape.reservation.reservation.ReservationRepository;
+import roomescape.reservation.reservation.ReservationStatus;
+import roomescape.reservation.reservation.dto.ReservationResponse;
+import roomescape.reservation.waiting.*;
 import roomescape.reservationtime.ReservationTime;
 import roomescape.reservationtime.ReservationTimeRepository;
 import roomescape.theme.Theme;
 import roomescape.theme.ThemeRepository;
-import roomescape.waiting.dto.WaitingRequest;
-import roomescape.waiting.dto.WaitingResponse;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
